+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.3.0 [unreleased]
 
+### Features
+1. [#32](https://github.com/influxdata/influxdb-client-ruby/pull/32): Checks the health of a running InfluxDB instance by querying the /health
+
 ## 1.2.0 [2020-03-13]
 
 ### Features